Forked memory_utils.hpp off utils.hpp.
Naming the file memory.hpp caused some files that #include "memory.hpp" to
include the wrong file on Windows, because another file in a subfolder
has that name.  So we name it memory_utils.hpp.
