+++ src/client_protocol/protocols.cc
@@ -3,6 +3,8 @@
 
 #include <limits>
 
+#include "utils.hpp"
+
 const uint32_t wire_protocol_t::HARD_LIMIT_TOO_LARGE_QUERY_SIZE = GIGABYTE;
 const uint32_t wire_protocol_t::TOO_LONG_QUERY_TIME = 5*60*1000; // ms
 const uint32_t wire_protocol_t::TOO_LARGE_QUERY_SIZE = 128 * MEGABYTE;
