+++ src/concurrency/pump_coro.cc
@@ -4,6 +4,7 @@
 #include "arch/runtime/coroutines.hpp"
 #include "concurrency/interruptor.hpp"
 #include "containers/map_sentries.hpp"
+#include "utils.hpp"
 
 pump_coro_t::pump_coro_t(
         const std::function<void(signal_t *)> &_callback,
@@ -105,7 +106,7 @@ void pump_coro_t::run(auto_drainer_t::lock_t keepalive) {
             break;
         }
     }
-    
+
     --running;
 }
