+++ src/utils.hpp
@@ -35,18 +35,6 @@ struct const_charslice {
     const_charslice() : beg(nullptr), end(nullptr) { }
 };
 
-void *raw_malloc_aligned(size_t size, size_t alignment);
-void raw_free_aligned(void *ptr);
-
-#ifndef _WIN32
-void *raw_malloc_page_aligned(size_t size);
-#endif
-
-/* Calls `malloc()` and checks its return value to crash if the allocation fails. */
-void *rmalloc(size_t size);
-/* Calls `realloc()` and checks its return value to crash if the allocation fails. */
-void *rrealloc(void *ptr, size_t size);
-
 /* Forwards to the isfinite macro, or std::isfinite. */
 bool risfinite(double);
